Using react router to handle analytics page views
